@@ -22,15 +22,15 @@ android {
         targetCompatibility = "8"
         sourceCompatibility = "8"
     }
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
 
     defaultConfig {
         applicationId "me.cutmail.disasterapp"
         minSdkVersion 26
-        targetSdkVersion 28
-        versionCode 19
-        versionName "1.9.0"
+        targetSdkVersion 29
+        versionCode 20
+        versionName "1.9.1"
     }
     buildTypes {
         release {
@@ -40,19 +40,19 @@ android {
 }
 
 dependencies {
-    implementation('com.crashlytics.sdk.android:crashlytics:2.10.0@aar') {
+    implementation('com.crashlytics.sdk.android:crashlytics:2.10.1@aar') {
         transitive = true
     }
-    implementation 'androidx.appcompat:appcompat:1.1.0-alpha05'
+    implementation 'androidx.appcompat:appcompat:1.1.0-rc01'
     implementation 'androidx.paging:paging-runtime:2.1.0'
     implementation 'com.jakewharton.timber:timber:4.7.1'
     implementation 'com.github.hotchemi:android-rate:1.0.1'
     implementation 'com.firebaseui:firebase-ui-firestore:4.3.2'
-    implementation 'com.google.firebase:firebase-database:17.0.0'
-    implementation 'com.google.firebase:firebase-core:16.0.9'
+    implementation 'com.google.firebase:firebase-database:19.0.0'
+    implementation 'com.google.firebase:firebase-core:17.2.0'
     implementation 'com.jakewharton:butterknife:10.1.0'
     annotationProcessor 'com.jakewharton:butterknife-compiler:10.1.0'
-    implementation 'com.google.android.gms:play-services-oss-licenses:16.0.2'
+    implementation 'com.google.android.gms:play-services-oss-licenses:17.0.0'
 }
 
 apply plugin: 'com.google.gms.google-services'
